Refuse to export if repository is not up to date
Although it already updated and patches is applied
in correct order, it's confusing as you will
get "invisible" commits on your next update.

This way, it's not possible to export without having the latest and
greates locally.

This way it's also easier to let it be up to the user to resolve any
conflicts that may arise, when trying to "export" on a outdated repository.